implementation of lua views

+++ userspace/libsinsp/filterchecks.cpp
@@ -3238,7 +3238,7 @@ char* sinsp_filter_check_reference::format_bytes(int64_t val, uint32_t str_len)
 					"%*" PRId64, str_len, val);
 	}
 
-	uint32_t len = strlen(m_getpropertystr_storage);
+	uint32_t len = (uint32_t)strlen(m_getpropertystr_storage);
 
 	if(len > str_len)
 	{
